api/requirements.txt
fastapi==0.75.0
hypercorn>=0.13.2
uvicorn
aiofiles==0.8.0
python-multipart==0.0.5
httpx>=0.22.0
pydantic==1.10.13

# DB stuff
mysqlclient==2.1.0
sqlmodel>=0.0.8
sqlalchemy>=1.4.35

cryptography
pyjwt
pyjwt[crypto]
unidecode